Move Gemma4 ISL sweep to tests/e2e and document in README
Extract run_demo_text() from text_demo_v2.py and relocate the
parametrized batch/long-context perf rows to tests/e2e/test_isl_sweep.py.
Update README with full-sweep and single-bucket run commands near the
Per-ISL benchmark table; adjust conftest wording for the new test path.
